- parsing MainActivity 로 이동완료. - FavoriteFragment getData() 이동 전.
